* Split spectroscopy and its analysis class.
