+++ qiskit_experiments/characterization/__init__.py
@@ -35,11 +35,12 @@ Analysis
 
     T1Analysis
     T2RamseyAnalysis
-    SpectroscopyAnalysis
+    ResonanceAnalysis
 """
 from .t1 import T1
 from .t1_analysis import T1Analysis
-from .qubit_spectroscopy import QubitSpectroscopy, SpectroscopyAnalysis
+from .qubit_spectroscopy import QubitSpectroscopy
+from .resonance_analysis import ResonanceAnalysis
 from .ef_spectroscopy import EFSpectroscopy
 from .t2ramsey import T2Ramsey
 from .t2ramsey_analysis import T2RamseyAnalysis
